LPS-53990 rename for consistency
