+++ portal-impl/test/integration/com/liferay/portal/kernel/display/context/bundle/basedisplaycontextfactory/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Description: Test Bundle for BaseDisplayContextProviderTest
+Bundle-Description: Test Bundle for BaseDisplayContextFactoryTest
 Bundle-Version: 1.0
 Private-Package: $(bundle.package).*
